microservicio_tts/requirements.txt
fastapi
uvicorn
openai-whisper
elevenlabs
speechify-api
python-dotenv
watson-nlp-runtime-client
